requirements.txt
kafka==1.3.5
pyyaml==3.12
requests==2.18.4